adding like functionality

+++ Exam Preparation/2. Exam/My Theater/src/views/details.js
@@ -1,4 +1,9 @@
 import { deleteTheater, getTheaterById } from "../api/data.js";
+import {
+  createLike,
+  getCountLikes,
+  getSpecificUserLike,
+} from "../api/likes.js";
 import { html, until, nothing } from "../lib.js";
 
 const detailsTemplate = (template) => html`
@@ -7,10 +12,18 @@ const detailsTemplate = (template) => html`
   </section>
 `;
 
-const eventTemplate = (theater, user, isOwner, onDelete) => html`
+const eventTemplate = (
+  theater,
+  user,
+  isOwner,
+  onDelete,
+  onLike,
+  userLike,
+  theaterLikes
+) => html`
   <div id="detailsBox">
     <div class="detailsInfo">
-      <h1>${theater.title}</h1>
+      <h1>Title: ${theater.title}</h1>
       <div>
         <img src="${theater.imageUrl}" />
       </div>
@@ -23,16 +36,24 @@ const eventTemplate = (theater, user, isOwner, onDelete) => html`
       <h4>Author: ${theater.author}</h4>
       <div class="buttons">
         ${isOwner
-          ? html`<a @click=${onDelete} class="btn-delete" href="/profile"
+          ? html`<a
+                @click=${onDelete}
+                class="btn-delete"
+                href="javascript:void(0)"
                 >Delete</a
               >
               <a class="btn-edit" href="/edit/${theater._id}">Edit</a>`
           : nothing}
-        ${user && !isOwner
-          ? html`<a class="btn-like" href="#">Like</a>`
+        ${user && !isOwner && userLike == 0
+          ? html`<a
+              @click=${() => onLike(theater._id)}
+              class="btn-like"
+              href="javascript:void(0)"
+              >Like</a
+            >`
           : nothing}
       </div>
-      <p class="likes">Likes: 0</p>
+      <p class="likes">Likes: ${theaterLikes}</p>
     </div>
   </div>
 `;
@@ -43,14 +64,41 @@ export const detailsPage = (ctx) => {
   ctx.render(detailsTemplate(getTheater()));
 
   async function getTheater() {
+    let userLike = null;
     const data = await getTheaterById(ctx.params.id);
+    if (user) {
+      userLike = await getSpecificUserLike(ctx.params.id, user.id);
+    }
+    const theaterLikes = await getCountLikes(ctx.params.id);
 
     const isOwner = user && data._ownerId == user.id;
 
-    return eventTemplate(data, user || false, isOwner, onDelete);
+    return eventTemplate(
+      data,
+      user || false,
+      isOwner,
+      onDelete,
+      onLike,
+      userLike,
+      theaterLikes
+    );
   }
 
-  async function onDelete() {
-    await deleteTheater(ctx.params.id);
+  async function onDelete(e) {
+    e.preventDefault();
+
+    const question = confirm("Are you sure you want to delete this event?");
+
+    if (question) {
+      await deleteTheater(ctx.params.id);
+
+      ctx.page.redirect("/profile");
+    }
+  }
+
+  async function onLike(theaterId) {
+    await createLike({ theaterId });
+
+    ctx.page.redirect(`/details/${theaterId}`);
   }
 };
